Factorial.java

+++ Factorial.java
@@ -8,7 +8,7 @@ public class Factorial
         Scanner s=new Scanner(System.in);
         System.out.println("Enter a number:");
         int n=s.nextInt();
-        int j=1;
+        long j=1;
         for (int i=1;i<=n;i++)
         {
             j=i*j;
